# target/generated-sources/jaxrs/com/warehouse/api/beans/WarehouseSearchResult.java
package com.warehouse.api.beans;

import java.util.ArrayList;
import java.util.List;
import javax.annotation.processing.Generated;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;

@JsonInclude(JsonInclude.Include.NON_NULL)
@JsonPropertyOrder({
    "items",
    "page",
    "pageSize",
    "totalElements"
})
@Generated("jsonschema2pojo")
public class WarehouseSearchResult {

    @JsonProperty("items")
    private List<Warehouse> items = new ArrayList<Warehouse>();
    @JsonProperty("page")
    private Integer page;
    @JsonProperty("pageSize")
    private Integer pageSize;
    @JsonProperty("totalElements")
    private Integer totalElements;

    @JsonProperty("items")
    public List<Warehouse> getItems() {
        return items;
    }

    @JsonProperty("items")
    public void setItems(List<Warehouse> items) {
        this.items = items;
    }

    @JsonProperty("page")
    public Integer getPage() {
        return page;
    }

    @JsonProperty("page")
    public void setPage(Integer page) {
        this.page = page;
    }

    @JsonProperty("pageSize")
    public Integer getPageSize() {
        return pageSize;
    }

    @JsonProperty("pageSize")
    public void setPageSize(Integer pageSize) {
        this.pageSize = pageSize;
    }

    @JsonProperty("totalElements")
    public Integer getTotalElements() {
        return totalElements;
    }

    @JsonProperty("totalElements")
    public void setTotalElements(Integer totalElements) {
        this.totalElements = totalElements;
    }

}